organize tests into read and write
